requirements.txt
Flask
flask-cors
gunicorn
Pillow
numpy
tensorflow
tf-keras-vis
scikit-imagereportlab